# marketing-contracts/src/main/resources/openapi/marketing-api.yaml
openapi: 3.1.0
info:
  title: Marketing Low-code Platform API
  version: 1.0.0
  description: R1 edge contract. Tenant and organization scope come from the verified identity in OIDC mode.
servers:
  - {url: http://localhost:8080, description: Local edge gateway}
security: [{oidc: []}]
...
paths:
...
  /api/v1/approvals/{caseId}/decisions:
    post:
      tags: [Control]
      operationId: decideApproval
      parameters: [{$ref: '#/components/parameters/CaseId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovalDecision'}
      responses: *ok
...
  /api/v1/compile:
    post: {tags: [Compiler], operationId: compileArtifact, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/artifacts/{artifactId}:
    get: {tags: [Compiler], operationId: getArtifact, parameters: [{$ref: '#/components/parameters/ArtifactId'}], responses: *ok}
...
  /api/v1/decisions:evaluate:
    post: {tags: [Decision], operationId: evaluateDecision, requestBody: {$ref: '#/components/requestBodies/Decision'}, responses: *ok}
  /api/v1/decisions:batch-evaluate:
    post:
      tags: [Decision]
      operationId: batchEvaluateDecision
      requestBody:
        required: true
        content: {application/json: {schema: {type: array, minItems: 1, maxItems: 50, items: {$ref: '#/components/schemas/DecisionRequest'}}}}
      responses: *list
  /api/v1/decisions/runtime/manifest:
    put: {tags: [Decision], operationId: warmDecisionRuntime, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/decisions/runtime/activation:
    put: {tags: [Decision], operationId: activateDecisionRuntime, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/decisions/runtime/ack:
    post: {tags: [Decision], operationId: refreshDecisionRuntimeAck, responses: *ok}
  /api/v1/decisions/runtime/audience-membership:
    put: {tags: [Decision], operationId: updateDecisionAudienceProjection, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
...
  /api/v1/experiments:
    post: {tags: [Measurement], operationId: createExperiment, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/experiments/{experimentId}/versions/{version}:assign:
    post: {tags: [Measurement], operationId: assignExperiment, parameters: [{$ref: '#/components/parameters/ExperimentId'}, {$ref: '#/components/parameters/StringVersion'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/experiments/{experimentId}/versions/{version}:check-srm:
    post: {tags: [Measurement], operationId: checkExperimentSrm, parameters: [{$ref: '#/components/parameters/ExperimentId'}, {$ref: '#/components/parameters/StringVersion'}], responses: *ok}
  /api/v1/measurements/facts:
    post: {tags: [Measurement], operationId: ingestMeasurementFact, requestBody: {$ref: '#/components/requestBodies/MarketingFact'}, responses: *accepted}
  /api/v1/measurements/watermarks:
    post: {tags: [Measurement], operationId: advanceMeasurementWatermark, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/measurements/dashboard:
    get:
      tags: [Measurement]
      operationId: getMeasurementDashboard
      parameters: [{in: query, name: from, required: true, schema: {type: string, format: date-time}}, {in: query, name: to, required: true, schema: {type: string, format: date-time}}]
      responses: *ok
  /api/v1/measurements/series:
    get:
      tags: [Measurement]
      operationId: getMeasurementSeries
      parameters:
        - {in: query, name: from, required: true, schema: {type: string, format: date-time}}
        - {in: query, name: to, required: true, schema: {type: string, format: date-time}}
        - {in: query, name: granularity, required: true, schema: {type: string, enum: [hour, day]}}
      responses: *ok
  /api/v1/measurements/attribution:recompute:
    post:
      tags: [Measurement]
      operationId: recomputeAttribution
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: *ok
  /api/v1/measurements/attribution/{conversionEventId}:
    get:
      tags: [Measurement]
      operationId: getAttribution
      parameters:
        - {in: path, name: conversionEventId, required: true, schema: {type: string}}
        - {in: query, name: policy, required: true, schema: {type: string, enum: [FIRST_TOUCH, LAST_TOUCH, LINEAR]}}
        - {in: query, name: windowSeconds, required: true, schema: {type: integer, format: int64, minimum: 1}}
      responses: *ok
  /api/v1/traces:
    post: {tags: [Measurement], operationId: recordDecisionTrace, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/traces/requests/{requestId}:
    get: {tags: [Measurement], operationId: getTraceByRequest, parameters: [{$ref: '#/components/parameters/RequestId'}], responses: *ok}
  /api/v1/traces/orders/{orderId}:
    get: {tags: [Measurement], operationId: getTraceByOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: *ok}
components:
  securitySchemes:
    oidc: {type: openIdConnect, openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration}
  parameters:
    IdempotencyKey: {in: header, name: Idempotency-Key, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
...
    StringVersion: {in: path, name: version, required: true, schema: {type: string}}
    CaseId: {in: path, name: caseId, required: true, schema: {type: string}}
...
    ArtifactId: {in: path, name: artifactId, required: true, schema: {type: string}}
...
    ExperimentId: {in: path, name: experimentId, required: true, schema: {type: string}}
    RequestId: {in: path, name: requestId, required: true, schema: {type: string}}
    OrderId: {in: path, name: orderId, required: true, schema: {type: string}}
  requestBodies:
    JsonObject: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/JsonObject'}}}}
...
    ApprovalDecision: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalDecision'}}}}
...
    Decision: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DecisionRequest'}}}}
...
    MarketingFact: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MeasurementFactRequest'}}}}
...
  schemas:
    JsonObject: {type: object, additionalProperties: true}
...
    ApprovalDecision:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {type: string, enum: [BUSINESS, FINANCE, COMPLIANCE, MERCHANT]}
        decision: {type: string, enum: [APPROVE, REJECT], default: APPROVE}
        comment: {type: string, maxLength: 128}
...
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [requestId, idempotencyKey, organizationId, subjectToken, channel, occurredAt, cart]
      properties:
        requestId: {type: string}
        idempotencyKey: {type: string}
        organizationId: {type: string}
        orderId: {type: string}
        subjectToken: {type: string}
        channel: {type: string}
        occurredAt: {type: string, format: date-time}
        cart: {$ref: '#/components/schemas/Cart'}
        audienceSnapshotIds: {type: array, uniqueItems: true, items: {type: string}}
    Cart:
      type: object
      additionalProperties: false
      required: [currency, lines]
      properties: {currency: {type: string, pattern: '^[A-Z]{3}$'}, lines: {type: array, minItems: 1, maxItems: 100, items: {$ref: '#/components/schemas/CartLine'}}}
    CartLine:
      type: object
      additionalProperties: false
      required: [lineId, skuId, shopId, quantity, unitPrice, floorUnitPrice, shippingLine]
      properties:
        lineId: {type: string}
        skuId: {type: string}
        shopId: {type: string}
        categoryId: {type: string}
        brandId: {type: string}
        quantity: {type: integer, minimum: 1, maximum: 100000}
        unitPrice: {$ref: '#/components/schemas/Money'}
        floorUnitPrice: {$ref: '#/components/schemas/Money'}
        shippingLine: {type: boolean}
    Money:
      type: object
      additionalProperties: false
      required: [currency, minorUnits]
      properties: {currency: {type: string, pattern: '^[A-Z]{3}$'}, minorUnits: {type: integer, format: int64, minimum: 0}}
...
    MeasurementFactRequest:
      type: object
      additionalProperties: false
      required: [eventId, type, businessKey, subjectToken, occurredAt, ingestedAt, schemaVersion, attributes, correctionOf]
      properties:
        eventId: {type: string}
        type: {type: string}
        businessKey: {type: string}
        subjectToken: {type: string}
        occurredAt: {type: string, format: date-time}
        ingestedAt: {type: string, format: date-time}
        schemaVersion: {type: string}
        attributes: {type: object, additionalProperties: {type: string}}
        correctionOf: {type: string}
...